fm_api_project/requirements.txt
appdirs==1.4.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-appconf==1.0.2
django-cors-headers==2.0.2
django-enumfield==1.2.1
django-extensions==1.7.9
django-imagekit==4.0.1
django-rest-auth==0.9.1
django-rest-authemail==1.4.0
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
httpie==0.9.9
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.0
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
python3-openid==3.1.0
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
stripe==1.62.0
uritemplate==3.0.0
urllib3==1.22
